add api and webui for configure parallelism, and revert default parallelism

+++ runtime_scan/pkg/config/scan_config.go
@@ -20,7 +20,7 @@ import (
 )
 
 type ScanConfig struct {
-	MaxScanParallelism           int
+	MaxScanParallelism           int64
 	TargetNamespaces             []string
 	IgnoredNamespaces            []string
 	JobResultTimeout             time.Duration
